requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.1.3
django-redis==5.4.0
djangorestframework==3.15.2
redis==5.2.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
